util: moving internal helpers to internal module
1. Moving `_errnoException` and `_exceptionWithHostPort` to
   `internal/util` module as they are internal helper functions. They
   should not be exposed as part of the `util` module.

2. Issuing a deprecation warning when those functions are used.
