+++ lib/child_process.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const util = require('util');
+const internalUtil = require('internal/util');
 const debug = util.debuglog('child_process');
 const constants = require('constants');
 
@@ -10,7 +11,7 @@ const Buffer = require('buffer').Buffer;
 const Pipe = process.binding('pipe_wrap').Pipe;
 const child_process = require('internal/child_process');
 
-const errnoException = util._errnoException;
+const errnoException = internalUtil._errnoException;
 const _validateStdio = child_process._validateStdio;
 const setupChannel = child_process.setupChannel;
 const ChildProcess = exports.ChildProcess = child_process.ChildProcess;
