+++ lib/dgram.js
@@ -3,6 +3,7 @@
 const assert = require('assert');
 const Buffer = require('buffer').Buffer;
 const util = require('util');
+const internalUtil = require('internal/util');
 const events = require('events');
 const constants = require('constants');
 
@@ -17,8 +18,8 @@ const BIND_STATE_BOUND = 2;
 var cluster = null;
 var dns = null;
 
-const errnoException = util._errnoException;
-const exceptionWithHostPort = util._exceptionWithHostPort;
+const errnoException = internalUtil._errnoException;
+const exceptionWithHostPort = internalUtil._exceptionWithHostPort;
 
 function lookup(address, family, callback) {
   if (!dns)
