+++ lib/dns.js
@@ -2,6 +2,7 @@
 
 const net = require('net');
 const util = require('util');
+const internalUtil = require('internal/util');
 
 const cares = process.binding('cares_wrap');
 const uv = process.binding('uv');
@@ -27,7 +28,7 @@ function errnoException(err, syscall, hostname) {
     ex.errno = err;
     ex.syscall = syscall;
   } else {
-    ex = util._errnoException(err, syscall);
+    ex = internalUtil._errnoException(err, syscall);
   }
   if (hostname) {
     ex.hostname = hostname;
