+++ lib/fs.js
@@ -5,6 +5,7 @@
 
 const SlowBuffer = require('buffer').SlowBuffer;
 const util = require('util');
+const internalUtil = require('internal/util');
 const pathModule = require('path');
 
 const binding = process.binding('fs');
@@ -34,7 +35,7 @@ const O_WRONLY = constants.O_WRONLY || 0;
 const isWindows = process.platform === 'win32';
 
 const DEBUG = process.env.NODE_DEBUG && /fs/.test(process.env.NODE_DEBUG);
-const errnoException = util._errnoException;
+const errnoException = internalUtil._errnoException;
 
 function throwOptionsError(options) {
   throw new TypeError('Expected options to be either an object or a string, ' +
