+++ lib/internal/child_process.js
@@ -6,6 +6,7 @@ const EventEmitter = require('events').EventEmitter;
 const net = require('net');
 const dgram = require('dgram');
 const util = require('util');
+const internalUtil = require('internal/util');
 const constants = require('constants');
 const assert = require('assert');
 
@@ -18,7 +19,7 @@ const TCP = process.binding('tcp_wrap').TCP;
 const UDP = process.binding('udp_wrap').UDP;
 const SocketList = require('internal/socket_list');
 
-const errnoException = util._errnoException;
+const errnoException = internalUtil._errnoException;
 const SocketListSend = SocketList.SocketListSend;
 const SocketListReceive = SocketList.SocketListReceive;
