+++ lib/internal/util.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const uv = process.binding('uv');
+
 exports.printDeprecationMessage = function(msg, warned) {
   if (process.noDeprecation)
     return true;
@@ -39,4 +41,43 @@ exports.deprecate = function(fn, msg) {
   }
 
   return deprecated;
+};
+
+exports._errnoException = function(err, syscall, original) {
+  const errname = uv.errname(err);
+  var message = syscall + ' ' + errname;
+
+  if (original)
+    message += ' ' + original;
+
+  const e = new Error(message);
+  e.code = errname;
+  e.errno = errname;
+  e.syscall = syscall;
+
+  return e;
+};
+
+
+exports._exceptionWithHostPort = function(err, syscall, address, port, extra) {
+  var details;
+
+  if (port && port > 0) {
+    details = address + ':' + port;
+  } else {
+    details = address;
+  }
+
+  if (extra) {
+    details += ' - Local (' + extra + ')';
+  }
+
+  const ex = exports._errnoException(err, syscall, details);
+
+  ex.address = address;
+  if (port) {
+    ex.port = port;
+  }
+
+  return ex;
 };
