+++ lib/net.js
@@ -4,6 +4,7 @@ const events = require('events');
 const stream = require('stream');
 const timers = require('timers');
 const util = require('util');
+const internalUtil = require('internal/util');
 const assert = require('assert');
 const cares = process.binding('cares_wrap');
 const uv = process.binding('uv');
@@ -19,8 +20,8 @@ const WriteWrap = process.binding('stream_wrap').WriteWrap;
 
 
 var cluster;
-const errnoException = util._errnoException;
-const exceptionWithHostPort = util._exceptionWithHostPort;
+const errnoException = internalUtil._errnoException;
+const exceptionWithHostPort = internalUtil._exceptionWithHostPort;
 
 function noop() {}
