+++ lib/tty.js
@@ -1,11 +1,12 @@
 'use strict';
 
 const util = require('util');
+const internalUtil = require('internal/util');
 const net = require('net');
 const TTY = process.binding('tty_wrap').TTY;
 const isTTY = process.binding('tty_wrap').isTTY;
 const inherits = util.inherits;
-const errnoException = util._errnoException;
+const errnoException = internalUtil._errnoException;
 
 
 exports.isatty = function(fd) {
