+++ lib/util.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const uv = process.binding('uv');
 const Buffer = require('buffer').Buffer;
 const Debug = require('vm').runInDebugContext('Debug');
 const internalUtil = require('internal/util');
@@ -806,38 +805,10 @@ exports.pump = exports.deprecate(function(readStream, writeStream, callback) {
 }, 'util.pump(): Use readableStream.pipe() instead');
 
 
-exports._errnoException = function(err, syscall, original) {
-  var errname = uv.errname(err);
-  var message = syscall + ' ' + errname;
-  if (original)
-    message += ' ' + original;
-  var e = new Error(message);
-  e.code = errname;
-  e.errno = errname;
-  e.syscall = syscall;
-  return e;
-};
+exports._errnoException = internalUtil.deprecate(
+  internalUtil._errnoException,
+  'util._errnoException is deprecated.');
 
-
-exports._exceptionWithHostPort = function(err,
-                                          syscall,
-                                          address,
-                                          port,
-                                          additional) {
-  var details;
-  if (port && port > 0) {
-    details = address + ':' + port;
-  } else {
-    details = address;
-  }
-
-  if (additional) {
-    details += ' - Local (' + additional + ')';
-  }
-  var ex = exports._errnoException(err, syscall, details);
-  ex.address = address;
-  if (port) {
-    ex.port = port;
-  }
-  return ex;
-};
+exports._exceptionWithHostPort = internalUtil.deprecate(
+  internalUtil._exceptionWithHostPort,
+  'util._exceptionWithHostPort is deprecated.');
